requirements.txt
torch>=2.2.0
torchvision>=0.17.0
tensorboard>=2.12
loguru>=0.7.2
PyYAML>=6.0.1
tqdm>=4.66.0
numpy>=1.24.0
datasets>=2.19.0
psutil>=5.9.0
rich>=13.7.0
